+++ lib/pages.dart
@@ -1,0 +1,1 @@
+export "./landingPage.dart";
