+++ lib/src/image_tools.dart
@@ -1,0 +1,20 @@
+import 'package:image/image.dart' as img;
+
+class ImageTools {
+  static img.Image invertImage(img.Image src, num shiftCoefficient) {
+    final pixels = src.getBytes();
+
+    int pixelId = -1;
+    while (pixelId < pixels.length - 1) {
+      pixelId += 1;
+      var pixel = pixels[pixelId];
+      if ((pixelId + 1) % 4 == 0) continue; // alpha channel
+      num newPixel = 255 - pixels[pixelId];
+      num difference = pixel - newPixel;
+      pixels[pixelId] =
+          (newPixel + difference * shiftCoefficient).clamp(0, 255).toInt();
+    }
+
+    return src;
+  }
+}
