+++ test/widget_test.dart
@@ -13,7 +13,7 @@ import 'package:picturevert/main.dart';
 void main() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
     // Build our app and trigger a frame.
-    await tester.pumpWidget(const MyApp());
+    await tester.pumpWidget(const PVApp());
 
     // Verify that our counter starts at 0.
     expect(find.text('0'), findsOneWidget);
